requirements.txt
openai>=1.40.0
python-dotenv>=1.0.0
python-evtx>=0.7
